% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\name{ingest_data}
\alias{ingest_data}
\title{Import all data from the MMD Tool}
\usage{
ingest_data(filepath)
}
\arguments{
\item{filepath}{filepath to MMD Collection Tool}
}
\description{
Import all data from the MMD Tool
}
