% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_month.R
\name{ingest_month}
\alias{ingest_month}
\title{Identify Month}
\usage{
ingest_month(path)
}
\arguments{
\item{path}{filepath to MMD Collection Tool}
}
\description{
Identify Month
}
\examples{
\dontrun{
#extract current month from tool
  path <- "~/MMDData/MMD National Data Collection Tool PEPFARlandia.xlsx"
  ingest_month(path) }
}
