% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_ou.R
\name{ingest_ou}
\alias{ingest_ou}
\title{Identify OU name}
\usage{
ingest_ou(path)
}
\arguments{
\item{path}{filepath to MMD Collection Tool}
}
\description{
Identify OU name
}
\examples{
\dontrun{
#structure output for Burundi
  path <- "~/MMDData/MMD National Data Collection Tool PEPFARlandia.xlsx"
  ingest_ou(path) }
}
