% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_regimens.R
\name{ingest_regimens}
\alias{ingest_regimens}
\title{Import Regimes Tab}
\usage{
ingest_regimens(filepath)
}
\arguments{
\item{filepath}{filepath to MMD Collection Tool}
}
\description{
Import Regimes Tab
}
