% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_sheetexist.R
\name{ingest_sheetexist}
\alias{ingest_sheetexist}
\title{Confirm that sheet exists}
\usage{
ingest_sheetexist(filepath, sheetname)
}
\arguments{
\item{filepath}{filepath to MMD Collection Tool}

\item{sheetname}{name of the sheet to check}
}
\description{
Confirm that sheet exists
}
