% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_txtargets.R
\name{ingest_txtargets}
\alias{ingest_txtargets}
\title{Import TX_CURR Targets}
\usage{
ingest_txtargets(filepath)
}
\arguments{
\item{filepath}{filepath to MMD Collection Tool}
}
\description{
Import TX_CURR Targets
}
