% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_version.R
\name{ingest_version}
\alias{ingest_version}
\title{Identify template version}
\usage{
ingest_version(path)
}
\arguments{
\item{path}{filepath to MMD Collection Tool}
}
\description{
Identify template version
}
\examples{
\dontrun{
#extract tool version
  path <- "~/MMDData/MMD National Data Collection Tool PEPFARlandia.xlsx"
  ingest_version(path) }
}
