% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_ou.R
\name{ingest_ou}
\alias{ingest_ou}
\title{Identify OU name}
\usage{
ingest_ou(filepath)
}
\arguments{
\item{filepath}{filepath to MMD Collection Tool}
}
\description{
Identify OU name
}
\examples{
\dontrun{
#pull ou name from tool
  filepath <- "~/MMDData/MMD National Data Collection Tool PEPFARlandia.xlsx"
  ingest_ou(filepath) }
}
